Fix up postgreSQL version check
Postgres version schema changed after 9.6

Need to handle version as a float to handle 9.6 and 11
This was an introduced bug in:
9561a91327acac18f7834d7424a1efed799b51ca

+++ trove/guestagent/datastore/experimental/postgresql/service.py
@@ -415,7 +415,7 @@ class PgSqlApp(object):
         """Wrapper for pg_current_xlog_location()
         Cannot be used against a running slave
         """
-        version = int(self.pg_version[1])
+        version = float(self.pg_version[1])
         if version < 10:
             query = "SELECT pg_current_xlog_location()"
         else:
@@ -427,7 +427,7 @@ class PgSqlApp(object):
         """Wrapper for pg_last_xlog_replay_location()
          For use on standby servers
          """
-        version = int(self.pg_version[1])
+        version = float(self.pg_version[1])
         if version < 10:
             query = "SELECT pg_last_xlog_replay_location()"
         else:
@@ -517,7 +517,7 @@ class PgSqlApp(object):
         return r[0][0]
 
     def pg_xlogfile_name(self, start_segment):
-        version = int(self.pg_version[1])
+        version = float(self.pg_version[1])
         if version < 10:
             query = "SELECT pg_xlogfile_name('%s')"
         else:
